Add files via upload
Research paper based on Raspberry PI 2 cluster: 12 nodes with one master node and 11 worker nodes. Each node has 32gb of memory via SD card and the master node is backed with an extra 320GB external hard drive. All nodes are hooked on a 16port ethernet switch limited to 100MBPS speed for any node. We will need to cite this in our work eventually, but this is a good basis to form our own cluster. We're going down a good path by making a Pi 4 as our master node.